Remove hacks and fix: The reasons for mult mismatch was because in filter function I had set up train_obs as actions instead of obs and actions was null. Fixed this.
Secondly observations had been constant, this is because obs was not updating with
new_obs (which wasn't getting reset with seed as well). Plus fixed some cosmetics.

The result is that we have a single instance of Net that is not re-worked upon,
and iterate_batches' action is actually calculated always from the incoming Net
with a softmax. Thus the feedback loop is built in and iterate_batches itself
becomes both training model as well as the inference engine.

The program nows runs without interruption. I ran it for a couple of iterate_batches
and generally see result_mean increasing. Let's see how long I need to run before
result_mean > 199 which would mean I have solved the cartpole problem using cross
entropy RL