Solved the cartpole problem using cross entropy deep RL
I eliminated all printfs (uncomment them later for debug and understanding)
and only kept the loss, reward_mean and reward_bound on to check if the
trend is converging. Remarkably I had convergence in 22 steps ! (the book cites 42)

Can see the big jump from 21 to 22. Visually it was a treat to see the cartpole
getting more and more stable (the agent travels in the right direction, and doesn't
do much if env itself is stable, i.e. it really learns to balance the cartpole.

Here is the first output and my feeling of elation in watching my first
RL algorithm run successfully on a primitive robot that learns to
balance an object.

python3 test2_cartpole.py
/Users/anupkaul/akaul_git/drl/simenvs/cross_entropy/test2_cartpole.py:117: UserWarning: Creating a tensor from a list of numpy.ndarrays is extremely slow. Please consider converting the list to a single numpy.ndarray with numpy.array() before converting to a tensor. (Triggered internally at /Users/runner/work/pytorch/pytorch/pytorch/torch/csrc/utils/tensor_new.cpp:264.)
  obs_v = torch.FloatTensor([obs])
0: loss=0.688, reward_mean=18.9, reward_bound=20.5
1: loss=0.669, reward_mean=34.3, reward_bound=45.0
2: loss=0.657, reward_mean=36.2, reward_bound=38.0
3: loss=0.642, reward_mean=50.9, reward_bound=59.5
4: loss=0.632, reward_mean=49.3, reward_bound=56.0
5: loss=0.627, reward_mean=43.8, reward_bound=52.0
6: loss=0.618, reward_mean=52.3, reward_bound=60.5
7: loss=0.601, reward_mean=64.3, reward_bound=72.0
8: loss=0.610, reward_mean=76.6, reward_bound=90.5
9: loss=0.579, reward_mean=63.4, reward_bound=65.0
10: loss=0.589, reward_mean=89.2, reward_bound=107.0
11: loss=0.590, reward_mean=93.0, reward_bound=106.0
12: loss=0.581, reward_mean=82.4, reward_bound=94.0
13: loss=0.581, reward_mean=101.9, reward_bound=104.5
14: loss=0.565, reward_mean=100.2, reward_bound=116.0
15: loss=0.551, reward_mean=107.1, reward_bound=129.0
16: loss=0.552, reward_mean=131.1, reward_bound=150.5
17: loss=0.554, reward_mean=135.8, reward_bound=133.5
18: loss=0.558, reward_mean=164.4, reward_bound=176.0
19: loss=0.549, reward_mean=166.2, reward_bound=183.0
20: loss=0.545, reward_mean=147.8, reward_bound=176.0
21: loss=0.556, reward_mean=168.5, reward_bound=185.5
22: loss=0.546, reward_mean=260.9, reward_bound=301.0
RL Solved !!