Extend play post RL solve for cartpole
The agent learns faster (mean reward jumps to 500 and above in 55-60 iterations)
Visually, gets much better around the 400-500 reward range. Agent balances better,
and there were some plays that were just beautiful. It learns "tricks", like moving
along with the pole, or move to the opposite side.

One artifact (maybe from the env?) is that the agent learns the trick of gliding past
the window to stop the stick from falling, vs trying to keep the stick always within
the widow.

1. Wonder if changing the env to include negative rewards when going out the window helps
2. Wonder what happens at 100 and above episodes? (this one was till 60)
3. Would like to preserve and reboot the agent with the last/best trained net.