Load model post saving and verify that it works for cartpole's cross entropy
View shows that multiple invocations of cartpole now resumes with mean_rewards
starting off at the last cumulative values ! (first time is 23.5, next start is
in the 40s and the 3rd is in the 80s). Can investigate if adding any additional
states along with the model can speed it up (using multiple dicts examples)

For the moment I have added a simple load command right after net is instantiated.
It seems to work