Figured that reducing the problem space may address frozen_lake cross entropy convergence
My 8** map was random and I'd have to wait for random successful goal reaches. Instead a
better approach is to go with the simplest default map first. Enabled this and logs show
that we are able to run the network now ! (elite batches/full batch is updated) and we
start generating loss values. Will need to run the network a bit now to get updates.

Also disabled loading the model as there were incompatibilities with the previous model.
May need to start a new cross entropy model from scratch, probably for the simple map.
Once I learn that model, can extend/udpdate it via RL on larger and random maps.