Quick hack to identify the dict in self.state tuple
(the dict doesn't in self.state doesn't allow me to use rewards
and transitions table, resulting in unhashable dict messages)

I still need to identify the 'name field' of this dict inside
self.state (the purpose of this field is to include the probability
of state when an action is taken. The clear part is obs expressed
as an int.) Have gone through the env code of Frozen Lake and haven't
landed on the name yet.

0, {'prob': 1})

 <class 'tuple'> 2

type:  <class 'int'> val:  0
type:  <class 'dict'> val:  {'prob': 1}