Solved self.state's type switching by overriding it to initial int everytime on env.reset
This works with original hashtables now for reward, transits and values table.
Code runs through with 100 plays and value iteration algorithm proceeds

PlayEpisodes with 20 episodes gets stuck. I'm guessing this is because we have
TWO environments. Not sure why test_env has to be different and instantiated outside
of the agent class, while the agent class has it's own env instance ..