Cleanup code (remove old prints), print self.rewards and self.transit values
We do iterate (still with random) and when accidentally the agent hits the target
best reward gets updated as well. We still need to figure out how to exit play_episode
if action is 0 (which it likely may be initially). Still random actions in play_episode

self.rewards and self.transit databases seem to be ok. Need to inspect self.value,
value_iter and the actual calc_action algorithm. Also will be interesting to store
the 3 collections because this is the disk representation of a network for this
example.